feat: Update Consumer#poll to raise an exception when message is an error
The message object returned by `LibRdKafka.consumer_poll` can be a
proper message or an event or error. In the case of an error we will now
raise a ConsumerException by default. The previous behaviour can be
maintained by passing `raise_on_error: false`.

Also increase timeouts in integration tests for Github CI.